requirements.txt
python-docx>=1.1.0
python-pptx>=0.6.23
openai>=1.0.0
customtkinter>=5.2.0
pyinstaller>=6.0.0
json_repair>=0.1.0
